+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Removed `settings.json`
+- Phylogenetic tree now made with react-bio-viz
 
 ### Fixed